Aggiunte metriche sul tempo di esecuzione

+++ src/encoder/Encoder.hpp
@@ -15,8 +15,8 @@ class Encoder {
         static void callback(int way);
     public:
         Encoder(int clk_gpio, int dt_gpio, int ppr, int start_angle_degrees);
-        float get_angle();
-        float get_angle_degrees();
+        double get_angle();
+        double get_angle_degrees();
 };
 
 #endif
